dataset-creator/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
fastuuid==0.14.0
filelock==3.20.1
frozenlist==1.8.0
fsspec==2025.12.0
future==1.0.0
grpcio==1.67.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
litellm==1.80.10
MarkupSafe==3.0.3
mido==1.3.3
multidict==6.7.0
numpy==2.3.5
openai==2.13.0
packaging==25.0
pedalboard==0.9.19
pretty_midi==0.2.11
propcache==0.4.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pylast==7.0.0
pyloudnorm==0.1.1
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.30.0
scipy==1.16.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
spotipy==2.25.2
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
yarl==1.22.0
zipp==3.23.0
